M2M100MT => M2M100
