+++ src/transformers/models/m2m_100/tokenization_m2m_100.py
@@ -1,4 +1,3 @@
-# coding=utf-8
 # Copyright Suraj Patil and The HuggingFace Inc. team. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,41 +11,32 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tokenization classes for M2M100MT."""
-import os
+"""Tokenization classes for M2M100."""
+import json
 from contextlib import contextmanager
+from pathlib import Path
 from shutil import copyfile
-from typing import List, Optional, Tuple
+from typing import Dict, List, Optional, Tuple, Union
 
-from tokenizers import processors
-from transformers.tokenization_utils_fast import PreTrainedTokenizerFast
+import sentencepiece
 
-from ...tokenization_utils import BatchEncoding
+from ...tokenization_utils import BatchEncoding, PreTrainedTokenizer
 from ...utils import logging
-from .tokenization_mtm_100_mt import M2M100MTTokenizer
 
 
 logger = logging.get_logger(__name__)
 
-PRETRAINED_VOCAB_FILES_MAP = {
-    "vocab_file": {
-        "m2m100_418M": "https://huggingface.co/m2m100_418M/resolve/main/vocab.json",
-    },
-    "merges_file": {
-        "m2m100_418M": "https://huggingface.co/m2m100_418M/resolve/main/merges.txt",
-    },
-    "tokenizer_file": {
-        "m2m100_418M": "https://huggingface.co/m2m100_418M/resolve/main/tokenizer.json",
-    },
-}
-
-PRETRAINED_POSITIONAL_EMBEDDINGS_SIZES = {
-    "m2m_100_418M": 1024,
-}
+SPIECE_UNDERLINE = "▁"
 
 VOCAB_FILES_NAMES = {
     "vocab_file": "vocab.json",
     "spm_file": "sentencepiece.bpe.model",
+}
+
+PRETRAINED_VOCAB_FILES_MAP = {
+    "vocab_file": {
+        "m2m_100_418M": "https://huggingface.co/m2m100_418M/resolve/main/vocab.json",
+    }
 }
 
 ALL_M2M100_MODELS = ["facebook/m2m100_418M", "facebook/m2m100_1.2B"]
@@ -57,13 +47,19 @@ FAIRSEQ_LANGUAGE_CODES = ["af", "am", "ar", "ast", "az", "ba", "be", "bg", "bn",
 # fmt: on
 
 
-class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
-    r"""
-    Construct a "fast" M2M100MT tokenizer (backed by HuggingFace's `tokenizers` library).
+class M2M100Tokenizer(PreTrainedTokenizer):
+    """
+    Construct an M2M100 tokenizer. Based on `SentencePiece <https://github.com/google/sentencepiece>`__.
 
-     Args:
+    This tokenizer inherits from :class:`~transformers.PreTrainedTokenizer` which contains most of the main methods.
+    Users should refer to this superclass for more information regarding those methods.
+
+    Args:
         vocab_file (:obj:`str`):
             Path to the vocabulary file.
+        spm_file (:obj:`str`):
+            Path to `SentencePiece <https://github.com/google/sentencepiece>`__ file (generally has a .spm extension)
+            that contains the vocabulary.
         src_lang (:obj:`str`, `optional`):
             A string representing the source language.
         tgt_lang (:obj:`str`, `optional`):
@@ -79,14 +75,11 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
             token instead.
         pad_token (:obj:`str`, `optional`, defaults to :obj:`"<pad>"`):
             The token used for padding, for example when batching sequences of different lengths.
-        mask_token (:obj:`str`, `optional`, defaults to :obj:`"<mask>"`):
-            The token used for masking values. This is the token used when training this model with masked language
-            modeling. This is the token which the model will try to predict.
 
     Examples::
 
-        >>> from transformers import M2M100MTTokenizerFast
-        >>> tokenizer = M2M100MTTokenizerFast.from_pretrained("facebook/m2m100_418M, src_lang="en_XX", tgt_lang="ro_RO")
+        >>> from transformers import M2M100Tokenizer
+        >>> tokenizer = M2M100Tokenizer.from_pretrained("facebook/m2m100_418M, src_lang="en_XX", tgt_lang="ro_RO")
         >>> src_text = " UN Chief Says There Is No Military Solution in Syria"
         >>> tgt_text =  "Şeful ONU declară că nu există o soluţie militară în Siria"
         >>> model_inputs = tokenizer(src_text, return_tensors="pt")
@@ -97,9 +90,8 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
 
     vocab_files_names = VOCAB_FILES_NAMES
     max_model_input_sizes = {m: 1024 for m in ALL_M2M100_MODELS}
-    pretrained_vocab_files_map = {"vocab_file": {m: SPM_URL for m in ALL_M2M100_MODELS}}
+    pretrained_vocab_files_map = PRETRAINED_VOCAB_FILES_MAP
     model_input_names = ["input_ids", "attention_mask"]
-    slow_tokenizer_class = M2M100MTTokenizer
 
     prefix_tokens: List[int] = []
     suffix_tokens: List[int] = []
@@ -107,20 +99,17 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
     def __init__(
         self,
         vocab_file,
+        spm_file,
         src_lang=None,
         tgt_lang=None,
-        tokenizer_file=None,
         bos_token="<s>",
         eos_token="</s>",
         sep_token="</s>",
-        unk_token="<unk>",
         pad_token="<pad>",
-        **kwargs
+        unk_token="<unk>",
+        **kwargs,
     ):
-
         super().__init__(
-            vocab_file,
-            tokenizer_file=tokenizer_file,
             bos_token=bos_token,
             eos_token=eos_token,
             sep_token=sep_token,
@@ -129,13 +118,20 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
             **kwargs,
         )
 
-        self.vocab_file = vocab_file
+        self.encoder = load_json(vocab_file)
+        self.decoder = {v: k for k, v in self.encoder.items()}
+        self.spm_file = spm_file
+        self.sp_model = load_spm(spm_file)
 
-        self.lang_tokens = [f"__{lang_code}__" for lang_code in FAIRSEQ_LANGUAGE_CODES]
-        self.add_special_tokens({"additional_special_tokens": self.lang_tokens})
+        self.encoder_size = len(self.encoder)
 
         self.lang_code_to_token = {lang_code: f"__{lang_code}__" for lang_code in FAIRSEQ_LANGUAGE_CODES}
-        self.lang_token_to_id = {lang_token: self.convert_tokens_to_ids(lang_token) for lang_token in self.lang_tokens}
+
+        self.lang_token_to_id = {
+            self.get_lang_token(lang_code): self.encoder_size + i for i, lang_code in enumerate(FAIRSEQ_LANGUAGE_CODES)
+        }
+        self.id_to_lang_token = {v: k for k, v in self.lang_token_to_id.items()}
+        self._additional_special_tokens = list(self.lang_token_to_id.keys())
 
         self._src_lang = src_lang if src_lang is not None else "en"
         self.tgt_lang = tgt_lang
@@ -146,10 +142,7 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
 
     @property
     def vocab_size(self) -> int:
-        """
-        :obj:`int`: Size of the base vocabulary (without the added tokens).
-        """
-        return self._tokenizer.get_vocab_size(with_added_tokens=False) + self.num_madeup_words
+        return len(self.encoder) + len(self.lang_token_to_id) + self.num_madeup_words
 
     @property
     def src_lang(self) -> str:
@@ -160,16 +153,35 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
         self._src_lang = new_src_lang
         self.set_src_lang_special_tokens(self._src_lang)
 
+    def _tokenize(self, text: str) -> List[str]:
+        return self.sp_model.EncodeAsPieces(text)
+
+    def _convert_token_to_id(self, token):
+        if token in self.lang_token_to_id:
+            return self.lang_token_to_id[token]
+        return self.encoder.get(token, self.encoder[self.unk_token])
+
+    def _convert_id_to_token(self, index: int) -> str:
+        """Converts an index (integer) in a token (str) using the decoder."""
+        if index in self.id_to_lang_token:
+            return self.id_to_lang_token[index]
+        return self.decoder.get(index, self.unk_token)
+
+    def convert_tokens_to_string(self, tokens: List[str]) -> str:
+        """Converts a sequence of tokens (strings for sub-words) in a single string."""
+        out_string = "".join(tokens).replace(SPIECE_UNDERLINE, " ").strip()
+        return out_string
+
     def get_special_tokens_mask(
         self, token_ids_0: List[int], token_ids_1: Optional[List[int]] = None, already_has_special_tokens: bool = False
     ) -> List[int]:
         """
-        Retrieves sequence ids from a token list that has no special tokens added. This method is called when adding
+        Retrieve sequence ids from a token list that has no special tokens added. This method is called when adding
         special tokens using the tokenizer ``prepare_for_model`` method.
 
         Args:
             token_ids_0 (:obj:`List[int]`):
-                List of ids.
+                List of IDs.
             token_ids_1 (:obj:`List[int]`, `optional`):
                 Optional second list of IDs for sequence pairs.
             already_has_special_tokens (:obj:`bool`, `optional`, defaults to :obj:`False`):
@@ -197,12 +209,10 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
     ) -> List[int]:
         """
         Build model inputs from a sequence or a pair of sequence for sequence classification tasks by concatenating and
-        adding special tokens. The special tokens depend on calling set_lang.
+        adding special tokens. An MBART sequence has the following format, where ``X`` represents the sequence:
 
-        An MBART-50 sequence has the following format, where ``X`` represents the sequence:
-
-        - ``input_ids`` (for encoder) ``[src_lang_code] X [eos]``
-        - ``labels``: (for decoder) ``[tgt_lang_code] X [eos]``
+        - ``input_ids`` (for encoder) ``X [eos, src_lang_code]``
+        - ``decoder_input_ids``: (for decoder) ``X [eos, tgt_lang_code]``
 
         BOS is never used. Pairs of sequences are not the expected use case, but they will be handled without a
         separator.
@@ -214,23 +224,55 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
                 Optional second list of IDs for sequence pairs.
 
         Returns:
-            :obj:`List[int]`: list of `input IDs <../glossary.html#input-ids>`__ with the appropriate special tokens.
+            :obj:`List[int]`: List of `input IDs <../glossary.html#input-ids>`__ with the appropriate special tokens.
         """
         if token_ids_1 is None:
             return self.prefix_tokens + token_ids_0 + self.suffix_tokens
         # We don't expect to process pairs, but leave the pair logic for API consistency
         return self.prefix_tokens + token_ids_0 + token_ids_1 + self.suffix_tokens
 
+    def get_vocab(self) -> Dict:
+        vocab = self.encoder.copy()
+        vocab.update(self.added_tokens_encoder)
+        return vocab
+
+    def __getstate__(self) -> Dict:
+        state = self.__dict__.copy()
+        state["sp_model"] = None
+        return state
+
+    def __setstate__(self, d: Dict) -> None:
+        self.__dict__ = d
+        self.sp_model = load_spm(self.spm_file)
+
+    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
+        save_dir = Path(save_directory)
+        assert save_dir.is_dir(), f"{save_directory} should be a directory"
+        vocab_save_path = save_dir / (
+            (filename_prefix + "-" if filename_prefix else "") + self.vocab_files_names["vocab_file"]
+        )
+        spm_save_path = save_dir / (
+            (filename_prefix + "-" if filename_prefix else "") + self.vocab_files_names["spm_file"]
+        )
+
+        save_json(self.encoder, vocab_save_path)
+
+        if not spm_save_path.exists():
+            copyfile(self.spm_file, spm_save_path)
+
+        return (str(vocab_save_path), str(spm_save_path))
+
     def prepare_seq2seq_batch(
         self,
         src_texts: List[str],
-        src_lang: str = "en_XX",
+        src_lang: str = "en",
         tgt_texts: Optional[List[str]] = None,
-        tgt_lang: str = "ro_RO",
+        tgt_lang: str = "ro",
         **kwargs,
     ) -> BatchEncoding:
         self.src_lang = src_lang
         self.tgt_lang = tgt_lang
+        self.set_src_lang_special_tokens(self.src_lang)
         return super().prepare_seq2seq_batch(src_texts, tgt_texts, **kwargs)
 
     @contextmanager
@@ -244,36 +286,18 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
         self.set_src_lang_special_tokens(self.src_lang)
 
     def set_src_lang_special_tokens(self, src_lang: str) -> None:
-        """Reset the special tokens to the source lang setting. prefix=[src_lang_code] and suffix=[eos]."""
+        """Reset the special tokens to the source lang setting. No prefix and suffix=[eos, src_lang_code]."""
         lang_token = self.get_lang_token(src_lang)
-        self.cur_lang_code_id = self.lang_token_to_id[lang_token]
-        self.prefix_tokens = [self.cur_lang_code_id]
+        self.cur_lang_id = self.lang_token_to_id[lang_token]
+        self.prefix_tokens = [self.cur_lang_id]
         self.suffix_tokens = [self.eos_token_id]
-
-        prefix_tokens_str = self.convert_ids_to_tokens(self.prefix_tokens)
-        suffix_tokens_str = self.convert_ids_to_tokens(self.suffix_tokens)
-
-        self._tokenizer.post_processor = processors.TemplateProcessing(
-            single=prefix_tokens_str + ["$A"] + suffix_tokens_str,
-            pair=prefix_tokens_str + ["$A", "$B"] + suffix_tokens_str,
-            special_tokens=list(zip(prefix_tokens_str + suffix_tokens_str, self.prefix_tokens + self.suffix_tokens)),
-        )
 
     def set_tgt_lang_special_tokens(self, tgt_lang: str) -> None:
-        """Reset the special tokens to the target language setting. prefix=[src_lang_code] and suffix=[eos]."""
+        """Reset the special tokens to the target language setting. No prefix and suffix=[eos, tgt_lang_code]."""
         lang_token = self.get_lang_token(tgt_lang)
         self.cur_lang_id = self.lang_token_to_id[lang_token]
-        self.prefix_tokens = [self.cur_lang_code_id]
+        self.prefix_tokens = [self.cur_lang_id]
         self.suffix_tokens = [self.eos_token_id]
-
-        prefix_tokens_str = self.convert_ids_to_tokens(self.prefix_tokens)
-        suffix_tokens_str = self.convert_ids_to_tokens(self.suffix_tokens)
-
-        self._tokenizer.post_processor = processors.TemplateProcessing(
-            single=prefix_tokens_str + ["$A"] + suffix_tokens_str,
-            pair=prefix_tokens_str + ["$A", "$B"] + suffix_tokens_str,
-            special_tokens=list(zip(prefix_tokens_str + suffix_tokens_str, self.prefix_tokens + self.suffix_tokens)),
-        )
 
     def get_lang_token(self, lang: str) -> str:
         return self.lang_code_to_token[lang]
@@ -282,15 +306,18 @@ class M2M100MTTokenizerFast(PreTrainedTokenizerFast):
         lang_token = self.get_lang_token(lang)
         return self.lang_token_to_id[lang_token]
 
-    def save_vocabulary(self, save_directory: str, filename_prefix: Optional[str] = None) -> Tuple[str]:
-        if not os.path.isdir(save_directory):
-            logger.error("Vocabulary path ({}) should be a directory".format(save_directory))
-            return
-        out_vocab_file = os.path.join(
-            save_directory, (filename_prefix + "-" if filename_prefix else "") + VOCAB_FILES_NAMES["vocab_file"]
-        )
 
-        if os.path.abspath(self.vocab_file) != os.path.abspath(out_vocab_file):
-            copyfile(self.vocab_file, out_vocab_file)
+def load_spm(path: str) -> sentencepiece.SentencePieceProcessor:
+    spm = sentencepiece.SentencePieceProcessor()
+    spm.Load(str(path))
+    return spm
 
-        return (out_vocab_file,)
+
+def load_json(path: str) -> Union[Dict, List]:
+    with open(path, "r") as f:
+        return json.load(f)
+
+
+def save_json(data, path: str) -> None:
+    with open(path, "w") as f:
+        json.dump(data, f, indent=2)
