+++ utils/check_repo.py
@@ -30,8 +30,8 @@ PATH_TO_DOC = "docs/source"
 # Being in this list is an exception and should **not** be the rule.
 IGNORE_NON_TESTED = [
     # models to ignore for not tested
-    "M2M100MTEncoder",  # Building part of bigger (tested) model.
-    "M2M100MTDecoder",  # Building part of bigger (tested) model.
+    "M2M100Encoder",  # Building part of bigger (tested) model.
+    "M2M100Decoder",  # Building part of bigger (tested) model.
     "LEDEncoder",  # Building part of bigger (tested) model.
     "LEDDecoder",  # Building part of bigger (tested) model.
     "BartDecoderWrapper",  # Building part of bigger (tested) model.
@@ -77,8 +77,8 @@ TEST_FILES_WITH_NO_COMMON_TESTS = [
 # should **not** be the rule.
 IGNORE_NON_AUTO_CONFIGURED = [
     # models to ignore for model xxx mapping
-    "M2M100MTEncoder",
-    "M2M100MTDecoder",
+    "M2M100Encoder",
+    "M2M100Decoder",
     "LEDEncoder",
     "LEDDecoder",
     "BartDecoder",
